app/requirements.txt
numpy
pandas
pymoo
psycopg2-binary
scikit-learn
matplotlib
